refactor: Performance and CRT effect to core.
