+++ src/core/PerformanceMode.ts
@@ -1,0 +1,6 @@
+export enum PerformanceMode {
+    AUTO = "auto",
+    HIGH = "high",
+    MEDIUM = "medium",
+    LOW = "low",
+}
